python/requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scipy>=1.10
